refactor(nested): :ok_hand: rename acc to accumulator
